+++ Assets/VektorVoxels/Voxels/VoxelData.cs
@@ -5,14 +5,11 @@ using VektorVoxels.Lighting;
 
 namespace VektorVoxels.Voxels {
     /// <summary>
-    /// Represents a single voxel stored voxel data instance.
-    /// Additional data that is both global and infrequently accessed is stored in the Voxel Table.
-    /// There may be more optimal data layouts but memory optimization is not a major focus of this project beyond basic
-    /// locality for meshing/lighting and alignment to make things easier on the GPU.
+    /// Represents a single voxel stored voxel instance.
     /// </summary>
     [StructLayout(LayoutKind.Sequential)]
     public readonly struct VoxelData : IEquatable<VoxelData> {
-        public readonly uint Id;          
+        public readonly ushort Id;          
         public readonly VoxelFlags Flags;
         public readonly FacingDirection Orientation;
         public readonly Color16 ColorData;
@@ -25,7 +22,7 @@ namespace VektorVoxels.Voxels {
         /// </summary>
         public static VoxelData Null() => new VoxelData(0, 0, 0, Color16.Clear());
 
-        public VoxelData(uint id, VoxelFlags flags, FacingDirection orientation, Color16 colorData) {
+        public VoxelData(ushort id, VoxelFlags flags, FacingDirection orientation, Color16 colorData) {
             Id = id;
             Flags = flags;
             Orientation = orientation;
